pyproject.toml
[project]
name = "biosignal-train"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "hydra-core",
  "omegaconf",
  "numpy",
  "pandas",
  "pyarrow",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy",
  "types-PyYAML",
  "pre-commit",
]
wandb = ["wandb"]
mlflow = ["mlflow"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # let formatter handle line length

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true  # pragmatic; tighten later

[project.scripts]
biosignals-train = "biosignals.cli.train:main"
biosignals-eval = "biosignals.cli.eval:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
  "integration: integration tests",
]
